requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-jazzmin==3.0.0
django-rest-passwordreset==1.4.2
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
idna==3.10
mysqlclient==2.2.4
pillow==10.4.0
PyJWT==1.7.1
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.3
